requirements.txt
Flask==3.0.0
Flask-Session==0.5.0
Werkzeug==3.0.1
Pillow==10.1.0
qrcode==7.4.2
reportlab==4.0.7
python-dotenv==1.0.0gunicorn==22.0.0